bloggers_project/bloggers_api/url.py
from django.conf.urls import url
from . import  views
from django.conf.urls import include
from rest_framework.routers import DefaultRouter

router = DefaultRouter()
router.register('profile', views.UserProfileViewSet)
router.register('login', views.LoginViewSet, base_name='login')
router.register('feed', views.UserProfileFeedViewSet)

urlpatterns =  [
                url(r'', include(router.urls))
]