requirements.txt
dj-database-url==0.5.0
Django==1.11
djangorestframework==3.6.2
gunicorn==19.9.0
pytz==2018.5
whitenoise==4.1
